feat: add token refresh handling for realtime console
